Cleanup and introduction of required properties
When using the archetype, the user now has to provide a version of the geoserver, a version of geotools and a name of the WPS project he wants to start. These values will be used in the pom.xml of the resulting WPS project. This pom.xml has also been cleaned up.